Move lru_cache implementation to separate files
